test(qwen35): tighten dflash evidence gates
